requirements.txt
pytest==7.4.3
PyYAML==6.0.1
typer==0.9.0